+++ db/migration/src/m021_clear_index_data.rs
@@ -101,7 +101,7 @@ impl MigrationTrait for Migration {
                     .col(ColumnDef::new(ChainInfo::SafeRegistryDST).binary_len(32).null())
                     .col(
                         ColumnDef::new(ChainInfo::MinIncomingTicketWinProb)
-                            .float()
+                            .double()
                             .not_null()
                             .default(1.0),
                     )
